fixed warning
git-svn-id: svn://jacobi.lnm.mw.tum.de/baci/trunk/ccarat@19814 060e22d8-9d1c-0410-868a-e3a950d63d95

+++ src/drt_acou/acou_visc_ele_calc.cpp
@@ -2469,7 +2469,7 @@ UpdateInteriorVariablesAndComputeResidual(DRT::Discretization &     discretizati
     tempVec1.Scale(dt*dirk_b[stage]/dirk_a[stage][stage]);
     hdgele->eleinteriorValnp_ += tempVec1;
 
-    if(unsigned(stage)==dirk_q-1)
+    if(stage==dirk_q-1)
       hdgele->eleinteriorValn_ = hdgele->eleinteriorValnp_;
   }
   else
